+++ Hyphenopoly_Loader.js
@@ -1,5 +1,5 @@
 /**
- * @license Hyphenopoly_Loader 2.8.1-devel - client side hyphenation
+ * @license Hyphenopoly_Loader 3.0.0 - client side hyphenation
  * ©2019  Mathias Nater, Zürich (mathiasnater at gmail dot com)
  * https://github.com/mnater/Hyphenopoly
  *