+++ CHANGELOG.md
@@ -2,17 +2,18 @@
 
 ## Unreleased
 ### Changed
-*   implement new hyphenation pattern binary (.hpb) format (issue #61)
+*   BREAKING CHANGE: implement new hyphenation pattern binary (.hpb) format (issue #61)
     *   update patterns to new format and include left-/rightmin according to source
     *   update hyphenEngine.asm/.wasm to consume new pattern format
     *   implement .hpb-version-check in Hyphenopoly.js and hyphenopoly.module.js
     *   fix issue #65
 
 ### Added
-*   tearDown event (issue #67)
+*   feat: tearDown event (issue #67)
+*   feat: loadError event (issue #59)
 
 ### Removed
-*   remove `Hyphenopoly.setup.classnames` – use [selectors](https://github.com/mnater/Hyphenopoly/wiki/Global-Hyphenopoly-Object#new-with-version-260-selectors) instead
+*   BREAKING CHANGE: remove `Hyphenopoly.setup.classnames` – use [selectors](https://github.com/mnater/Hyphenopoly/wiki/Global-Hyphenopoly-Object#new-with-version-260-selectors) instead
 
 ## Version 2.8.0 (Feb 28, 2019)
 *   the error event now accepts a `lvl` field ("info"/"warn"/"error") and logs accordingly (issue #56)